Fix three data races with the indicator thread
The three "success" global variables are written by the main thread and
read by the indicator thread, constituting a data race. Simplest option
is to make them atomics, as no further synchronization is truly needed
to avoid a race condition.

A cast loads the values for passing to printf. This is an "int" cast
despite the underlying types because that is the type expected for the
format string (i.e. "%i"). The original types were incorrect, and the
compiler warning about this is suppressed due to indirection through
replaceColors().

There is still both a data race and race condition on output_length, to
be addressed separately.

+++ src/clipboard.cpp
@@ -428,7 +428,7 @@ void setupIndicator(std::stop_token st) {
         }
     } else if ((action == Action::PipeIn || action == Action::PipeOut) && stderr_is_tty) {
         for (int i = 0; !st.stop_requested(); i == 9 ? i = 0 : i++) {
-            output_length = fprintf(stderr, replaceColors(working_message).data(), doing_action[action].data(), bytes_success, "B", spinner_steps.at(i).data());
+            output_length = fprintf(stderr, replaceColors(working_message).data(), doing_action[action].data(), (int)bytes_success, "B", spinner_steps.at(i).data());
             fflush(stderr);
             cv.wait_for(lock, std::chrono::milliseconds(50), [&]{ return st.stop_requested(); });
         }
@@ -760,7 +760,7 @@ void showFailures() {
 
 void showSuccesses() {
     if (action == Action::PipeIn || action == Action::PipeOut && stderr_is_tty) {
-        fprintf(stderr, replaceColors(pipe_success_message).data(), did_action[action].data(), bytes_success);
+        fprintf(stderr, replaceColors(pipe_success_message).data(), did_action[action].data(), (int)bytes_success);
         return;
     }
     if ((files_success == 1 && directories_success == 0) || (files_success == 0 && directories_success == 1)) {
@@ -771,11 +771,11 @@ void showSuccesses() {
         }
     } else {
         if ((files_success > 1) && (directories_success == 0)) {
-            printf(replaceColors(multiple_files_success_message).data(), did_action[action].data(), files_success);
+            printf(replaceColors(multiple_files_success_message).data(), did_action[action].data(), (int)files_success);
         } else if ((files_success == 0) && (directories_success > 1)) {
-            printf(replaceColors(multiple_directories_success_message).data(), did_action[action].data(), directories_success);
+            printf(replaceColors(multiple_directories_success_message).data(), did_action[action].data(), (int)directories_success);
         } else if ((files_success >= 1) && (directories_success >= 1)) {
-            printf(replaceColors(multiple_files_directories_success_message).data(), did_action[action].data(), files_success, directories_success);
+            printf(replaceColors(multiple_files_directories_success_message).data(), did_action[action].data(), (int)files_success, (int)directories_success);
         }
     }
 }
